requirements.txt
msgpack-python
pyzmq
tweepy
lxml
pika

